+++ shared/src/main/scala/co/topl/shared/InputOutput.scala
@@ -25,7 +25,6 @@ case class MintingStatusRequest(
 case class MintingStatusResponse(
     mintingStatus: String,
     address: String,
-    bridgePKey: String,
     redeemScript: String
 )
 
@@ -37,7 +36,9 @@ case class StartPeginSessionResponse(
     sessionID: String,
     script: String,
     escrowAddress: String,
-    descriptor: String
+    descriptor: String,
+    minHeight: Long,
+    maxHeight: Long
 )
 
 case class StartPegoutSessionResponse(