+++ topl-btc-bridge/src/test/scala/co/topl/bridge/managers/SessionManagerSpec.scala
@@ -16,8 +16,9 @@ class SessionManagerSpec extends CatsEffectSuite {
     "redeemAddress",
     "escrowAddress",
     "scriptAsm",
-    "toplBridgePKey",
     "sha256",
+    1,
+    100,
     "claimAddress",
     PeginSessionState.PeginSessionStateWaitingForBTC
   )