+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Logging all command line arguments when starting the application.
 - Support for sad path: User did not send the funds.
 - Support for sad path: Bridge did not mint the tokens in time.
+- Support for sad path: User did not redeem the tokens in time.
+- Add new parameter `--topl-blocks-to-recover` to specify the number of blocks to wait for the Topl network to recover.
+- Add two new fields to the `StartPeginSessionResponse` message: `minHeight` and `maxHeight`.
+This is needed to know the range of blocks to wait for the redemption.
+
 
 ### Changed
 
@@ -29,8 +34,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Front-end to use the new minting status.
 - No longer hardcodes the user's public key. User can provide it during session creation
 - Session creation also displays the escrow script for the user
+- Rename `--blocks-to-recover` to `--btc-blocks-to-recover`.
+- Update the contract for the redemption to include the height. The new contract
+`threshold(1, sha256($sha256) and height($min, $max))`.
 
 ### Removed
 
 - `/api/confirm-deposit-btc` WS was removed.
 - `/api/confirm-redemption` WS was removed.
+- Removed the signature from the contract for the redemption.
+- Removed the bridge public key from the `MintingStatusResponse`, as it is not
+needed anymore.